+++ recipes/deploy-admin.rb
@@ -22,6 +22,11 @@ region = node.fetch(:region, 'us-east-1')
 s3_file_archive_bucket_name = get_s3_file_archive_bucket_name
 # S3 file archive service
 
+# IBM Watson service credentials
+ibm_watson_user_info = get_ibm_watson_user_info
+ibm_watson_username = ibm_watson_user_info[:user]
+ibm_watson_psw = ibm_watson_user_info[:pass]
+
 capture_agent_monitor_url = node.fetch(
   :capture_agent_monitor_url, 'http://example.com/monitor_url'
 )
@@ -97,7 +102,7 @@ deploy_revision "matterhorn" do
     # Admin Specific
     install_otherpubs_service_config(most_recent_deploy, matterhorn_repo_root, auth_host)
     install_aws_s3_file_archive_service_config(most_recent_deploy, region, s3_file_archive_bucket_name)
-    # /Admin Specific
+    install_ibm_watson_transcription_service_config(most_recent_deploy, ibm_watson_username, ibm_watson_psw)
     install_published_event_details_email(most_recent_deploy, public_engage_hostname)
     configure_newrelic(most_recent_deploy, newrelic_app_name, :admin)